Move SpecializedGetters to org.apache.spark.sql.catalyst.data.
